cleanup

+++ image/moments.hpp
@@ -28,23 +28,23 @@ public:
      * \param j1
      * \return
      */
-    double getM(int i, int j)
+    long double getM(int i, int j)
     {
         QPair<int,int> c(i,j);
         if (!moment_m.contains(c))
         {
-            double ret = 0;
+            long double ret = 0;
             if (!m_mtx.isEmpty())
             {
-                QVector<double> xp, yp;
-                for (quint32 k=0; k<m_mtx.width(); ++k)  xp << qPow((double)k/(double)(m_mtx.width ()-1),i);
-                for (quint32 l=0; l<m_mtx.height(); ++l) yp << qPow((double)l/(double)(m_mtx.height()-1),j);
+                QVector<long double> xp, yp;
+                for (quint32 k=0; k<m_mtx.width(); ++k)  xp << qPow((long double)k/(long double)(m_mtx.width ()-1),i);
+                for (quint32 l=0; l<m_mtx.height(); ++l) yp << qPow((long double)l/(long double)(m_mtx.height()-1),j);
 
                 for (quint32 l=0; l<m_mtx.height(); ++l)
                 {
-                    double lsum = 0;
+                    long double lsum = 0;
                     T* line = m_mtx.line(l);
-                    for (quint32 k=0; k<m_mtx.width(); ++k) if (line[k]!=0.0) lsum += (xp[k]*(double)line[k]);
+                    for (quint32 k=0; k<m_mtx.width(); ++k) if (line[k]!=0.0) lsum += (xp[k]*(long double)line[k]);
                     lsum*=yp[l];
                     ret+=lsum;
                 }
@@ -57,13 +57,13 @@ public:
                 if (m_v.isEmpty()) foreach(p, m_pos)
                 {
                     const int x = p.first, y = p.second;
-                    ret = ret + (double)(qPow(x,i)*qPow(y,j));
+                    ret = ret + (long double)(qPow(x,i)*qPow(y,j));
                 }
                 else foreach(p, m_pos)
                 {
                     ++k;
                     const int x = p.first, y = p.second;
-                    ret = ret + (double)m_v[k] * (double)(qPow(x,i)*qPow(y,j));
+                    ret = ret + (long double)m_v[k] * (long double)(qPow(x,i)*qPow(y,j));
                 }
             }
             moment_m[c]=ret;
@@ -77,23 +77,23 @@ public:
      * \param j
      * \return
      */
-    double getMu(int i, int j)
+    long double getMu(int i, int j)
     {
         QPair<int,int> c(i,j);
         if (!moment_mu.contains(c))
         {
             if ((i==0) && (j==0)) return getM(0,0);
-            double xm = xMean(), ym = yMean(), ret = 0;
+            long double xm = xMean(), ym = yMean(), ret = 0;
             if (!m_mtx.isEmpty())
             {
-                QVector<double> xp, yp;
-                for (quint32 k=0; k<m_mtx.width (); ++k) xp << qPow(((double)k/(double)(m_mtx.width ()-1))-xm,i);
-                for (quint32 l=0; l<m_mtx.height(); ++l) yp << qPow(((double)l/(double)(m_mtx.height()-1))-ym,j);
+                QVector<long double> xp, yp;
+                for (quint32 k=0; k<m_mtx.width (); ++k) xp << qPow(((long double)k/(long double)(m_mtx.width ()-1))-xm,i);
+                for (quint32 l=0; l<m_mtx.height(); ++l) yp << qPow(((long double)l/(long double)(m_mtx.height()-1))-ym,j);
                 for (quint32 l=0; l<m_mtx.height(); ++l)
                 {
-                   double lsum = 0;
+                    long double lsum = 0;
                     T* line = m_mtx.line(l);
-                    for (quint32 k=0; k<m_mtx.width(); ++k) if (line[k]!=0.0) lsum += xp[k]*(double)line[k];
+                    for (quint32 k=0; k<m_mtx.width(); ++k) if (line[k]!=0.0) lsum += xp[k]*(long double)line[k];
                     lsum*=yp[l];
                     ret+=lsum;
                 }
@@ -105,14 +105,14 @@ public:
                 int k = -1;
                 if (m_v.isEmpty()) foreach(p, m_pos)
                 {
-                    const double x = p.first, y = p.second;
-                    ret = ret + (double)(qPow((x-xm),i)*qPow((y-ym),j));
+                    const long double x = p.first, y = p.second;
+                    ret = ret + (long double)(qPow((x-xm),i)*qPow((y-ym),j));
                 }
                 else foreach(p, m_pos)
                 {
                     ++k;
-                    const double x = p.first, y = p.second;
-                    ret = ret + (double)m_v[k] * (double)(qPow((x-xm),i)*qPow((y-ym),j));
+                    const long double x = p.first, y = p.second;
+                    ret = ret + (long double)m_v[k] * (long double)(qPow((x-xm),i)*qPow((y-ym),j));
                 }
             }
             moment_mu[c]=ret;
@@ -126,13 +126,13 @@ public:
      * \param j
      * \return
      */
-    double getEta(int i, int j)
+    long double getEta(int i, int j)
     {
         Q_ASSERT(i+j>=2);
         QPair<int,int> c(i,j);
         if (!moment_eta.contains(c))
         {
-            const double pwr = 2+i+j, ret = getMu(i,j) / qPow(getMu(0,0), pwr*0.5);
+            const long double pwr = 2+i+j, ret = getMu(i,j) / qPow(getMu(0,0), pwr*0.5);
             moment_eta[c]=ret;
         }
         return moment_eta[c];
@@ -143,7 +143,7 @@ public:
      * \param i 1..7
      * \return
      */
-    double getHu(int i)
+    long double getHu(int i)
 {
     if (!m_hu.contains(i)) switch(i)
     {
@@ -161,7 +161,7 @@ public:
         break;
         case 5:
         {
-            const double    // (30-312)(30+12)
+            const long double    // (30-312)(30+12)
                             d1 = (getEta(3,0)-3.0*getEta(1,2))*(getEta(3,0)+getEta(1,2)),
                             // (30+12)^2-3(21+03)^2
                             d2 = qPow(getEta(3,0)+getEta(1,2),2.0)-3.0*qPow(getEta(2,1)+getEta(0,3),2.0),
@@ -174,7 +174,7 @@ public:
         break;
         case 6:
         {
-            const double    // (20-02)
+            const long double    // (20-02)
                             d1 = getEta(2,0)-getEta(0,2),
                             // (30+12)^2-(21+03)^2
                             d2 = qPow(getEta(3,0)+getEta(1,2),2.0)-qPow(getEta(2,1)+getEta(0,3),2.0),
@@ -185,8 +185,9 @@ public:
         break;
         case 7:
         {
-            const double
+            const long double
                     //(321-03)(30+12)
+
                     d1 = (3.0*getEta(2,1)-getEta(0,3)) * (getEta(3,0)+getEta(1,2)),
                     //(30+12)^2-3(21+03)^2
                     d2 = qPow(getEta(3,0)+getEta(1,2),2.0)-3.0*qPow(getEta(2,1)+getEta(0,3),2.0),
@@ -209,14 +210,13 @@ public:
      * \param i
      * \return
      */
-    double getNHu(int i)
+    long double getNHu(int i)
     {
         if (!m_nhu.contains(i))
         {
-            const double hu = getHu(i);
+            const long double hu = getHu(i);
             if (hu==0.0) return 0.0;
             m_nhu[i] = (hu>0) ? -log(qAbs(hu)) : log(qAbs(hu));
-
         }
         return m_nhu[i];
     }
@@ -225,22 +225,24 @@ public:
     {
         QStringList sl;
         QPair<int,int> p;
-        for (int i=1; i<8; ++i)        sl << QString("Hu  #%1: %2").arg(i).arg(getHu(i));
-        for (int i=1; i<8; ++i)        sl << QString("NHu #%1: %2").arg(i).arg(getNHu(i));
-                                       sl << QString("Centroid %1 %2").arg(xMean()).arg(yMean());
-        foreach (p, moment_m.  keys()) sl << QString("M  %1%2: %3").arg(p.first).arg(p.second).arg(moment_m[p]);
-        foreach (p, moment_mu. keys()) sl << QString("Mu %1%2: %3").arg(p.first).arg(p.second).arg(moment_mu[p]);
-        foreach (p, moment_eta.keys()) sl << QString("Eta%1%2: %3").arg(p.first).arg(p.second).arg(moment_eta[p]);
+        for (int i=1; i<8; ++i)        sl << QString("Hu  #%1: %2").arg(i).arg((double)getHu(i));
+        for (int i=1; i<8; ++i)        sl << QString("NHu #%1: %2").arg(i).arg((double)getNHu(i));
+
+        sl << QString("Centroid %1 %2").arg((double)xMean()).arg((double)yMean());
+
+        foreach (p, moment_m.  keys()) sl << QString("M  %1%2: %3").arg(p.first).arg(p.second).arg((double)moment_m[p]);
+        foreach (p, moment_mu. keys()) sl << QString("Mu %1%2: %3").arg(p.first).arg(p.second).arg((double)moment_mu[p]);
+        foreach (p, moment_eta.keys()) sl << QString("Eta%1%2: %3").arg(p.first).arg(p.second).arg((double)moment_eta[p]);
         foreach(const QString& s, sl) qWarning("%s", qPrintable(s));
     }
 
 
 private:
-    inline double xMean() { const double m00 = getM(0,0); return (m00!=0) ? getM(1,0)/m00 : 0.0; }
-    inline double yMean() { const double m00 = getM(0,0); return (m00!=0) ? getM(0,1)/m00 : 0.0; }
+    inline long double xMean() { const long double m00 = getM(0,0); return (m00!=0) ? getM(1,0)/m00 : 0.0; }
+    inline long double yMean() { const long double m00 = getM(0,0); return (m00!=0) ? getM(0,1)/m00 : 0.0; }
 
-    QMap<QPair<int,int>,double> moment_m, moment_mu, moment_eta;
-    QMap<int,double> m_hu, m_nhu;
+    QMap<QPair<int,int>,long double> moment_m, moment_mu, moment_eta;
+    QMap<int,long double> m_hu, m_nhu;
     SScMatrix<T> m_mtx;
     QVector<QPair<int,int> > m_pos;
     QVector<T> m_v;
